third_party/blink/web_tests/external/wpt/offscreen-canvas/text/2d.text.draw.space.basic.worker-expected.txt

This is a testharness.js-based test.
FAIL U+0020 is rendered the correct size (1em wide) assert_approx_equals: Red channel of the pixel at (25, 25) expected 0 +/- 2 but got 255
Harness: the test ran to completion.
